src/requirements.txt
# Python Dependencies for Azure AI Agent Framework MVP
# Phase 2: Agent Implementation

# Microsoft Agent Framework (Preview - Nov 2025)
agent-framework>=1.0.0

# Azure OpenAI & Identity
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0

# Bot Framework SDK for Python
botbuilder-core>=4.15.0
botbuilder-schema>=4.15.0
aiohttp>=3.9.0

# FastAPI & Web Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Application Insights Telemetry
opencensus-ext-azure>=1.1.13
opencensus-ext-logging>=0.1.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# MCP Integration (Phase 3)
httpx>=0.25.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
